[RFC] nvme: make spdk_nvme_ctrlr_connect_io_qpair() non-blocking
In order to make `spdk_nvme_ctrlr_connect_io_qpair()` non-blocking, the
the polling was moved to `spdk_nvme_qpair_process_completions()`.

Additionally, the condition on which the requests are queued was
adjusted to make sure we never queue the connect fabrics command.
This allows other IO requests to be sent before the connection is
completed.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I04e45b1c2f49f9da3412c843ea899341c56a0420