nvme: qpair connected callback
Now that qpair connection can be asynchronous, users don't know when the
qpair gets connected and whether the connection was successful.  Since
this information can be very important to some users (e.g. NVMe
multipath needs to keep track of the state of a qpair), a callback was
added to send a notification when a qpair gets connected.

The callback is specified when creating a poll group: once a qpair that
is added to the poll group becomes connected, the callback will be
executed.  Tying the callback to poll groups doesn't require changes to
the qpair APIs, which is very mature at this point, as well as makes it
more consistent with the disconnect callback, which is also tied to poll
groups.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ibfae468dd6558c720c3612e3baa88637e3e8e441