sock/uring: asynchronous readv support
This patch adds the support for asynchronous reads in the io_uring sock
module.  The reads are performed directly to user's buffers, unless
there's some data in the recv_pipe at the time an asynchronous read is
submitted.  In that case, the data is copied from the pipe and either an
async read is scheduled or user's callback is executed immediately,
depending on whether request's buffer was fully filled.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I1c23a05038b6135a78e36b27c78bbd26752ac8d8