[RFC] sock/uring: async recv to the pipe from sock_group_poll
Reads to the recv_pipe are now submitted using the io_uring interface
during sock_group_poll().  If the user wants to read something from the
socket using the synchronous interface, spdk_sock_(readv|recv)(), the
data is only read from the pipe and a syscall is never sent from this
context.

Of course, if a socket is not part of a sock group or doesn't have a
recv_pipe allocated, the existing path will be used.

Currently, we try to fill the whole pipe that way.  It might be worth to
change that in the future, so that the pipe is filled up to some defined
threshold, but from the initial tests it didn't seem to have any
significant difference.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ic0089861296d0ace344a3ab9c49e0e69a3cc5c01