nvmf/tcp: don't reset recv PDU when setting qpair error state
This is done in preparation for reading PDU payload asynchronously.  To
send the async payload read, we'll be using pdu.sock_req as the request.
So, if we have an outstanding read request and encounter an error that
should put the qpair in NVME_TCP_PDU_RECV_STATE_ERROR state, we cannot
memset() the pdu_in_progress, as that would also memset the pdu.sock_req
still in use by the sock layer.

Therefore, we skip the memset() for NVME_TCP_PDU_RECV_STATE_ERROR.  And
a qpair should never transition out of this state (an assertion was also
added to verify that), so that memset() shouldn't be necessary.

There used to be one place where this was done (handling
nvmf_tcp_req_parse_sgl() failure), but that wasn't necessary, as the
qpair is also set to NVME_TCP_PDU_RECV_STATE_AWAIT_PDU_READY which also
clears the PDU.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I5b2f251ae83fa35eaaca4c307d23ab0d39f6e6c9