python: add setup script
It allows people to install our python scripts as a regular python
package via pip.  For now the script is very basic, but we can extend it
in the future (e.g. list dependencies).

Also, python uses a different version format than we do (see PEP 440),
so the version reported by the python scripts will be a little
different: the leading zero in the minor part is removed and "-pre" is
replaced with "rc0".  For instance, "23.05-pre" will be reported as
"23.5rc0".

Additionally, a simple test was added that will make sure that the
version in the scripts is consistent with the version in the .c code.

Fixes #2997.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I86b70adb121bf824ccabb6968103b7f6ffc3150d