nvme: don't pack spdk_nvme_ns_cmd_ext_io_opts
We're often using structure packing throughout the code for the *_opts
structs to make sure their size will always change whenever a new field
is added.  However, packing a structure has a couple of drawbacks:
__attribute__((packed)) is a compiler extension and isn't supported on
some compilers (e.g. msvc), and taking the address of a member field
might lead to Waddress-of-packed-member warnings.

Instead, pad the structure manually and use a new macro, SPDK_SIZEOF(),
to calculate the size of the structure.  The macro takes two parameters:
pointer to a structure and the name of the last member in that
structure.  It returns the size of the structure up to and including
that member.  That way, even though a structure isn't packed, the size
calculated using SPDK_SIZEOF() won't include the padding at the end.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I049baf4db0cd08d7cc728e31eda2ee6bad47b264