bdev: track IOs with accel sequence with an extra flag
We already have u.bdev.accel_sequence and internal.accel_sequence, but
both of them can be cleared in different cases.  The former is cleared
when a bdev doesn't support accel or by the module itself to notify bdev
layer that it executed the sequence.  The latter is cleared before
submitting a request to a module supporting accel, so that bdev layer
won't touch it again.

However, there are cases where it's necessary to know whether an IO was
associated with an accel sequence and a module has executed it (e.g. to
allow modules support accel without supporting memory domains), so an
extra flag is required.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I58ec789cce84b63151d02e28a0bb3d3bbf85f7bf