test/bdev: verify chaining integration with NVMe/TCP
Add an extra case to the chaining test that runs rw traffic with
data digest enabled and verifies that the total number of executed
operations/sequences is as expected.

Currently, the TCP transport doesn't report support for accel sequences,
so the operations cannot be chained and the expected number of executed
accel sequences is equal to the total number of operations.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ie4bad7ffc9b221dc1a008c4664fb3f652c480767