lib: add ut library
The purpose of this new library is to provide helpers functions for
common unit test tasks.  In the first iteration, it'll provide an
interface for running the tests based on the command-line parameters.
To use it, the calls to CU_basic_run_tests() need to be replaced with
spdk_ut_run_tests().

When working on a feature, it's often useful to run a single test case,
as it makes easier to read the logs.  So, the command-line interface
provides parameters to select the test case (-t, --test) or test suite
(-s, --suite) to execute.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ia8fd4bc047e7c6adf21a71215a0fd3d4f06bc834