test/unit: use spdk_ut_run_tests()
Replaced direct calls to the CUnit's functions to run the tests with
spdk_ut_run_tests().  That way, each test will have the ability to run
a specific test case.

The blob.c unit test wasn't changed, because it runs all tests multiple
times with different parameter combinations, so it cannot be easily
converted.  In the future, each such combination could be split into a
separate test suite, which would make it possible to convert it to using
spdk_ut_run_tests().

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I4463f808f89844e9bf32b5b31eda197c5d729d1d