nvmf/tcp: fix data digest error reporting
When a data digest error is detected, only the status.sc field of the
CQE was filled, leaving sct and cid unassigned.  The former isn't a big
deal, because the CQE is always zeroed when a new request arrives and
incidentally sct=GENERIC is zero.  However, not filling cid means that
the target would always report data digest errors for cid=0, regardless
of the actual id of the command.

Change-Id: Ia6d7be365cc9416472cd4c4cc479bb695552fdb4
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>