lib: add keyring library
The purpose of this library is to provide other libraries with an
interface for retrieving cryptographic keys.  This interface consists of
functions managing the keyring (add/remove), keeping track of key
references (get/put), and a set of callbacks used to define modules
responsible for providing access to the keys in a secure manner.

Key objects (called spdk_key) are identified by a unique name.  They
don't hold the keying material themselves, but only act as references to
the actual key.  Once the key is needed to perform a crypto operation,
this reference is used to retrieve keying material.  These objects are
refcounted, so it is safe to remove a key from the keyring while it's
still in use.  If that happens, the spdk_key object remains valid, but
it cannot be used to obtain the key (i.e. spdk_key_get_key() will result
in -ENOKEY).

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I5324fa28d5828a989b003bb5f715cfcac2077262