lib: add keyring library
The purpose of this library is to provide other libraries with an
interface for retrieving cryptographic keys.  This interface consists of
functions managing the keyring (add/remove), keeping track of key
references (get/put), and a set of callbacks used to define modules
responsible for providing access to the keys in a secure manner.

Key objects (called spdk_key) are identified by a unique name.  They
don't hold the keying material themselves, but only act as references to
the actual key.  Once user needs the key for crypto operations, this
reference can be used to retrieve keying material.  These objects are
refcounted, so it's safe to remove a key from the keyring, even if it's
still in use.  If that happens, the spdk_key objects are still valid,
but it's not possible to use them to get a key (i.e. spdk_key_get_key()
will result in -ENOKEY).

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I5324fa28d5828a989b003bb5f715cfcac2077262