check_so_deps: use make to get a list of deps from lib_deps.mk
Not only is this cleaner, but it also takes into account the
configuration, i.e. if a dependency is conditionally added based on one
of the CONFIG_* variables, it'll only be included if that variable is
set.

It makes it a lot easier to run this test locally, as it's no longer
necessary to enable all dependencies in configure.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ib5a809576dd1fe3d3a251a0dd021377b0f576cdc