keyring: add probe_key
This callbacks makes it possible to implement keyring modules that
instantiate the keys only when they're requested by the user.  For
instance, a module can fetch the key from a remote server over the
network.

Since probed keys are added to the keyring implicitly, they're removed
once the last reference to that key is released.  Users can still
explicitly remove such keys while they're still in use - in this case
the references are still valid, but they cannot be used to retrieve
the keying material.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I4852c8371c076bdf0f6cfce16cd86619ac4b8c7a