module/keyring: add linux keyring module
This module uses the keyutils library to retrieve keys managed by the
Linux kernel keyring.  The keys are created during the probe_key()
callback and, for now, they need to be attached to the kernel's keyring
before that.  The following patches will add the ability to specify
callout info passed to the kernel allowing it to instantiate the keys if
they're unavailable.

Change-Id: I0aedd19247c6c35f22bcb61f4b387d5d85b42cdf
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>