[WIP] keyring/linux: add callout info
It's now possible to specify callout info that will be passed to the
kernel in request_key(2).  If probing for a key that isn't already in
the kernel keyring, the kernel will try to instantiate it by executing
the /sbin/request-key program which will then run user-defined
application and pass it that callout info.  More details on the
configuration of that program can be found in request-key(5) man page.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I0b1737d0449087f3ba0e23fc23073d5612e7cb8c